+++ test/integration/component/component-exchange-event-call-on.js
@@ -131,8 +131,8 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: MESH_NAME_1 mesh does not exist on the api',
-          'invalid endpoint options: MESH_NAME_1 mesh does not exist on the api'
+          'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api',
+          'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api'
         ]);
     });
 
@@ -150,8 +150,8 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: MESH_NAME.component3 component does not exist on the api',
-          'invalid endpoint options: MESH_NAME.component3 component does not exist on the api'
+          'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api',
+          'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api'
         ]);
     });
 
@@ -167,8 +167,8 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: component3 component does not exist on the api',
-          'invalid endpoint options: component3 component does not exist on the api'
+          'invalid endpoint options: [component3] component does not exist on the api',
+          'invalid endpoint options: [component3] component does not exist on the api'
         ]);
     });
 
@@ -340,8 +340,8 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: MESH_NAME_1 mesh does not exist on the api',
-          'invalid endpoint options: MESH_NAME_1 mesh does not exist on the api'
+          'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api',
+          'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api'
         ]);
     });
 
@@ -359,8 +359,8 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: MESH_NAME.component3 component does not exist on the api',
-          'invalid endpoint options: MESH_NAME.component3 component does not exist on the api'
+          'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api',
+          'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api'
         ]);
     });
 
@@ -376,8 +376,25 @@ describe(tests.testName(__filename, 3), function() {
       tests
         .expect(errors)
         .to.eql([
-          'invalid endpoint options: component3 component does not exist on the api',
-          'invalid endpoint options: component3 component does not exist on the api'
+          'invalid endpoint options: [component3] component does not exist on the api',
+          'invalid endpoint options: [component3] component does not exist on the api'
+        ]);
+    });
+
+    it('we get errors attempting to access a method that does not exist', async () => {
+      let errors = await callWithInvalidParameters(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'unknownMethod'
+        },
+        { component: 'component', path: 'test/event' }
+      );
+      tests
+        .expect(errors)
+        .to.eql([
+          'invalid endpoint options: [component.unknownMethod] method does not exist on the api'
         ]);
     });
   });
@@ -421,6 +438,218 @@ describe(tests.testName(__filename, 3), function() {
       }
       tests.expect(message).to.be(undefined);
       tests.expect(result).to.be(3);
+    });
+  });
+
+  context('client tests - using callback', () => {
+    it('we are able to call component methods and listen to events - with callback', function(done) {
+      callAndListenCallback(
+        clientAdmin,
+        {
+          component: 'component',
+          method: 'exec'
+        },
+        { component: 'component1', path: 'test/event' },
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql({
+            event: { result: 3 },
+            exec: 3
+          });
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events, with mesh name - with callback', function(done) {
+      callAndListenCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql({
+            event: { result: 3 },
+            exec: 3
+          });
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events using $once, with mesh name - with callback', function(done) {
+      callAndListenOnceCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql(1);
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events with an $off, with mesh name - with callback', function(done) {
+      callAndListenOffCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        false,
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql(3);
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events with an $off, with mesh name - with callback - negative test', function(done) {
+      callAndListenOffCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        true,
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql(4);
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events with an $offPath, with mesh name - with callback', function(done) {
+      callAndListenOffPathCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        false,
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql(2);
+          done();
+        }
+      );
+    });
+
+    it('we are able to call component methods and listen to events with an $offPath, with mesh name - with callback - negative test', function(done) {
+      callAndListenOffPathCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component1', path: 'test/event' },
+        true,
+        (e, results) => {
+          if (e) return done(e);
+          tests.expect(results).to.eql(4);
+          done();
+        }
+      );
+    });
+
+    it('we get errors attempting to access a mesh that does not exist - with callback', function(done) {
+      callWithInvalidParametersCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME_1',
+          component: 'component',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME_1', component: 'component1', path: 'test/event' },
+        (e, errors) => {
+          tests
+            .expect(errors)
+            .to.eql([
+              'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api',
+              'invalid endpoint options: [MESH_NAME_1] mesh does not exist on the api'
+            ]);
+          done();
+        }
+      );
+    });
+
+    it('we get errors attempting to access a component that does not exist on a specified mesh - with callback', function(done) {
+      callWithInvalidParametersCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component3',
+          method: 'exec'
+        },
+        { mesh: 'MESH_NAME', component: 'component3', path: 'test/event' },
+        (e, errors) => {
+          tests
+            .expect(errors)
+            .to.eql([
+              'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api',
+              'invalid endpoint options: [MESH_NAME.component3] component does not exist on the api'
+            ]);
+          done();
+        }
+      );
+    });
+
+    it('we get errors attempting to access a component that does not exist - with callback', function(done) {
+      callWithInvalidParametersCallback(
+        clientAdmin,
+        {
+          component: 'component3',
+          method: 'exec'
+        },
+        { component: 'component3', path: 'test/event' },
+        (e, errors) => {
+          tests
+            .expect(errors)
+            .to.eql([
+              'invalid endpoint options: [component3] component does not exist on the api',
+              'invalid endpoint options: [component3] component does not exist on the api'
+            ]);
+          done();
+        }
+      );
+    });
+
+    it('we get errors attempting to access a method that does not exist - with callback', function(done) {
+      callWithInvalidParametersCallback(
+        clientAdmin,
+        {
+          mesh: 'MESH_NAME',
+          component: 'component',
+          method: 'unknownMethod'
+        },
+        { component: 'component', path: 'test/event' },
+        (e, errors) => {
+          tests
+            .expect(errors)
+            .to.eql([
+              'Error: invalid endpoint options: [component.unknownMethod] method does not exist on the api'
+            ]);
+          done();
+        }
+      );
     });
   });
 
@@ -498,6 +727,162 @@ describe(tests.testName(__filename, 3), function() {
     await client.exchange.$call(callParameters);
     await tests.delay(2000);
     return eventCounter;
+  }
+
+  function callWithInvalidParametersCallback(client, callParameters, listenParameters, callback) {
+    let errors = [];
+
+    client.event.$on(
+      listenParameters,
+      function() {
+        //do nothing
+      },
+      e => {
+        if (e) errors.push(e.message);
+        client.exchange.$call(callParameters, e => {
+          if (e) errors.push(e.message);
+          callback(null, errors);
+        });
+      }
+    );
+  }
+
+  function callAndListenCallback(client, callParameters, listenParameters, callback) {
+    let results = {};
+    client.event.$on(
+      listenParameters,
+      function(data) {
+        results.event = data;
+      },
+      e => {
+        if (e) return callback(e);
+        client.exchange.$call(callParameters, (e, result) => {
+          if (e) return callback(e);
+          results.exec = result;
+          setTimeout(() => {
+            callback(null, results);
+          }, 2000);
+        });
+      }
+    );
+  }
+
+  function callAndListenOnceCallback(client, callParameters, listenParameters, callback) {
+    let eventCounter = 0;
+    client.event.$once(
+      listenParameters,
+      function() {
+        eventCounter++;
+      },
+      e => {
+        if (e) return callback(e);
+        client.exchange.$call(callParameters, e => {
+          if (e) return callback(e);
+          client.exchange.$call(callParameters, e => {
+            if (e) return callback(e);
+            client.exchange.$call(callParameters, e => {
+              if (e) return callback(e);
+              setTimeout(() => {
+                callback(null, eventCounter);
+              }, 2000);
+            });
+          });
+        });
+      }
+    );
+  }
+
+  function callAndListenOffCallback(client, callParameters, listenParameters, negative, callback) {
+    let eventCounter = 0;
+    let id;
+
+    let finishCallback = () => {
+      client.exchange.$call(callParameters, e => {
+        if (e) return callback(e);
+        setTimeout(() => {
+          callback(null, eventCounter);
+        }, 2000);
+      });
+    };
+
+    client.event.$on(
+      listenParameters,
+      function() {
+        eventCounter++;
+      },
+      (e, eventId) => {
+        if (e) return callback(e);
+        id = eventId;
+        client.event.$on(
+          listenParameters,
+          function() {
+            eventCounter++;
+          },
+          e => {
+            if (e) return callback(e);
+            client.exchange.$call(callParameters, e => {
+              if (e) return callback(e);
+              if (negative) return finishCallback();
+              return client.event.$off(
+                {
+                  component: listenParameters.component,
+                  mesh: listenParameters.mesh,
+                  id
+                },
+                e => {
+                  if (e) return callback(e);
+                  finishCallback();
+                }
+              );
+            });
+          }
+        );
+      }
+    );
+  }
+
+  function callAndListenOffPathCallback(
+    client,
+    callParameters,
+    listenParameters,
+    negative,
+    callback
+  ) {
+    let eventCounter = 0;
+    let finishCallback = () => {
+      client.exchange.$call(callParameters, e => {
+        if (e) return callback(e);
+        setTimeout(() => {
+          callback(null, eventCounter);
+        }, 2000);
+      });
+    };
+    client.event.$on(
+      listenParameters,
+      function() {
+        eventCounter++;
+      },
+      e => {
+        if (e) return callback(e);
+        client.event.$on(
+          listenParameters,
+          function() {
+            eventCounter++;
+          },
+          e => {
+            if (e) return callback(e);
+            client.exchange.$call(callParameters, e => {
+              if (e) return callback(e);
+              if (negative) return finishCallback();
+              client.event.$offPath(listenParameters, e => {
+                if (e) return callback(e);
+                finishCallback();
+              });
+            });
+          }
+        );
+      }
+    );
   }
 
   async function setUpTestUser() {