+++ RELEASES.md
@@ -729,5 +729,13 @@ based on happner 1.28.1
   - SMC-1645: updated releases, docs and package
   - SMC-1645: refactor removed strange branch logic event and exchange layer
 
+11.6.1 2021-03-01
+-----------------
+SMC-1645: added first callback test
+SMC-1645: updated the internals to allow for callbacks on $ api
+SMC-1645: fixed and tested callback on unknown method, tested unknown method error on async
+SMC-1645: made error messages more consistent, encapsulating mesh.component.method in []'
 
-
+11.6.2 2021-03-01
+-----------------
+- SMC-2013: move sinon to devdeps